% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cpal.R
\docType{data}
\name{cpal}
\alias{cpal}
\title{List of Colour Palettes}
\format{\preformatted{List of 3
 $ seqtob: chr [1:12] "#000000" "#300828" "#631237" "#8d2531" ...
 $ quacla: chr [1:6] "#D23737" "#3191C9" "#D2BC2D" "#4EC93B" ...
 $ quagoo: chr [1:7] "#EC5549" "#797EF8" "#DCD22D" "#2CDA60" ...
}}
\usage{
cpal
}
\description{
This list speciefies a variety of colour palettes in hex-values.
}
\details{
Palette name: \code{<seq|qua|div><3_letter_palette_name>}
}
\keyword{datasets}

